IUCN Species Data 2024

A dataset containing the IUCN Red List conservation status of various species, including subspecies and subpopulations. The data includes species scientific names, threat categories, and additional information such as the core ID and rank.

Usage

iucn_2024

Format

A tibble with 166,503 rows and 7 columns:

category

character. The IUCN Red List category for each species, indicating its risk of extinction. Categories include "CR" (Critically Endangered), "EN" (Endangered), "VU" (Vulnerable), "LC" (Least Concern), "NT" (Near Threatened), "DD" (Data Deficient), and others.

threat_status

character. A descriptive representation of the conservation status, corresponding to the IUCN Red List category, e.g., "Critically Endangered", "Vulnerable", "Least Concern".

core_id

integer. A unique identifier for each species or subspecies within the dataset.

scientific_name

character. The scientific name of the species, including the genus and species epithet.

subspecies

character. The subspecies name, if applicable.

rank

character. The taxonomic rank below species level, such as "var." (variety) or "subsp." (subspecies).

subpopulation

character. The specific subpopulation designation, if applicable, for species that are divided into distinct subpopulations.

Source

International Union for Conservation of Nature (IUCN) Red List 2024.

Details

This dataset provides the conservation status as determined by the IUCN Red List for 2024. The data includes species with subspecies and subpopulation-level classifications where available. Subpopulations are identified for some species, and specific subpopulations are assigned threat statuses separately.

Examples

# Load the data
data(iucn_2024)

# View the first few rows of the dataset
head(iucn_2024)
#> # A tibble: 6 × 7
#>   category threat_status  core_id scientific_name subspecies rank  subpopulation
#>   <chr>    <chr>            <int> <chr>           <chr>      <chr> <chr>        
#> 1 DD       Data Deficient  1.31e8 Aa calceata     NA         NA    NA           
#> 2 DD       Data Deficient  1.31e8 Aa mandonii     NA         NA    NA           
#> 3 DD       Data Deficient  1.31e8 Aa matthewsii   NA         NA    NA           
#> 4 DD       Data Deficient  2.05e8 Abarema ferrug… NA         NA    NA           
#> 5 DD       Data Deficient  1.67e5 Abbottina binhi NA         NA    NA           
#> 6 DD       Data Deficient  4.26e4 Abditomys lati… NA         NA    NA