IUCN Red List Taxon Data

This dataset contains taxonomic information from the IUCN Red List of Threatened Species, downloaded from the GBIF website. It includes detailed taxonomic classifications and the conservation status of species across various kingdoms, offering a comprehensive view of the species listed in the Red List.

Usage

iucn_data

Format

A tibble with 150,490 rows and 21 variables:

search_name

character. A search-friendly version of the species' scientific name.

core_id

character. A unique identifier linking taxon data with distribution and threat status data.

scientific_name

character. The full scientific name of the taxon, including authorship.

kingdom

character. The kingdom classification of the taxon (e.g., PLANTAE).

phylum

character. The phylum classification of the taxon.

class

character. The class of the taxon.

order

character. The taxonomic order of the taxon.

family

character. The taxonomic family to which the taxon belongs.

genus

character. The genus to which the taxon belongs.

specific_epithet

character. The specific epithet of the taxon, forming the species name along with the genus.

scientific_name_authorship

character. The author(s) who described the taxon.

taxon_rank

character. The rank of the taxon (e.g., species, subspecies, variety).

infraspecific_epithet

character. The epithet for infraspecific taxa, such as subspecies or varieties, if applicable.

infrasp_tag

character. Indicator of the rank for infraspecific taxa (e.g., subspecies, variety, forma).

taxonomic_status

character. The taxon's status, indicating whether it is an accepted name or a synonym.

accepted_name_usage_id

integer. The identifier of the accepted name for synonyms, or the same as the core ID for accepted names.

bibliographic_citation

character. The citation for the taxon entry in the IUCN Red List.

references

character. A URL link to the IUCN Red List entry for the taxon.

source

character. The source of the distribution data.

threat_status

character. The IUCN conservation status of the species (e.g., "Endangered", "Least Concern", "Vulnerable").

occurrence_status

character. Indicates whether the taxon is present or absent in a given locality (e.g., "Present", "Absent").

The data is sourced from two primary files:

taxon.txt: Provides taxonomic information, including scientific names, taxon ranks, and accepted name usage IDs.

distribution.txt: Contains information about the IUCN threat status and occurrence for each taxon.

Source

The IUCN Red List of Threatened Species, available through GBIF: https://www.gbif.org/dataset/19491596-35ae-4a91-9a98-85cf505f1bd3

Examples

# Load the data
data(iucn_data)

# View the first few rows of the dataset
head(iucn_data)
#> # A tibble: 6 × 21
#>   search_name    core_id scientific_name kingdom phylum class order family genus
#>   <chr>          <chr>   <chr>           <chr>   <chr>  <chr> <chr> <chr>  <chr>
#> 1 Calocedrus ru… 133722  Calocedrus rup… PLANTAE TRACH… PINO… PINA… CUPRE… Calo…
#> 2 Narthecium re… 157938  Narthecium rev… PLANTAE TRACH… LILI… DIOS… NARTH… Nart…
#> 3 Asparagus hum… 158019  Asparagus humi… PLANTAE TRACH… LILI… ASPA… ASPAR… Aspa…
#> 4 Bulbostylis b… 158236  Bulbostylis bo… PLANTAE TRACH… LILI… POAL… CYPER… Bulb…
#> 5 Bulbostylis h… 158275  Bulbostylis hi… PLANTAE TRACH… LILI… POAL… CYPER… Bulb…
#> 6 Ophrys melite… 162229  Ophrys meliten… PLANTAE TRACH… LILI… ASPA… ORCHI… Ophr…
#> # ℹ 12 more variables: specific_epithet <chr>,
#> #   scientific_name_authorship <chr>, taxon_rank <chr>,
#> #   infraspecific_epithet <chr>, taxonomic_status <chr>,
#> #   accepted_name_usage_id <int>, bibliographic_citation <chr>,
#> #   references <chr>, source <chr>, threat_status <chr>,
#> #   occurrence_status <chr>, infrasp_tag <chr>